Only require file creator where its needed.
