+++ template.rb
@@ -1,11 +1,9 @@
 require_relative './gem_configurator.rb'
 require_relative './gemfile_configurator.rb'
-require_relative './file_remover.rb'
 require_relative './file_modifier.rb'
 require_relative './database_setup.rb'
 require_relative './installer.rb'
 require_relative './initialize_git.rb'
-require_relative './file_creator.rb'
 
 # -----------------------------
 # CREATE TEMPLATE
